% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigKnn.R
\docType{package}
\name{BigKnn-package}
\alias{BigKnn}
\alias{BigKnn-package}
\title{BigKnn: Large Scale K-Nearest Neighbor Classifier using the Lucene Search Engine}
\description{
A large scale k-nearest neighbor classifier using the Lucene search engine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/BigKnn}
  \item \url{https://github.com/OHDSI/BigKnn}
  \item Report bugs at \url{https://github.com/OHDSI/BigKnn/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Jenna Reps \email{jreps@its.jnj.com}
}

}
\keyword{internal}
