% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlpInterface.R
\name{buildKnnFromPlpData}
\alias{buildKnnFromPlpData}
\title{Build a K-nearest neighbor (KNN) classifier from a plpData object}
\usage{
buildKnnFromPlpData(
  plpData,
  population,
  indexFolder,
  overwrite = TRUE,
  cohortId = NULL,
  outcomeId = NULL
)
}
\arguments{
\item{plpData}{An object of type \code{plpData}.}

\item{population}{The population.}

\item{indexFolder}{Path to a local folder where the KNN classifier index can be stored.}

\item{overwrite}{Automatically overwrite if an index already exists?}

\item{cohortId}{The ID of the specific cohort for which to fit a model.}

\item{outcomeId}{The ID of the specific outcome for which to fit a model.}
}
\description{
Build a K-nearest neighbor (KNN) classifier from a plpData object
}
