% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildClassifier.R
\name{buildKnn}
\alias{buildKnn}
\title{Build a K-nearest neighbor (KNN) classifier}
\usage{
buildKnn(outcomes, covariates, indexFolder, overWrite = TRUE,
  checkSorting = TRUE, checkRowIds = TRUE, quiet = FALSE)
}
\arguments{
\item{outcomes}{A ffdf object containing the outcomes with predefined columns (see below).}

\item{covariates}{A ffdf object containing the covariates with predefined columns (see below).}

\item{indexFolder}{Path to a local folder where the KNN classifier index can be stored.}

\item{checkSorting}{Check if the data are sorted appropriately, and if not, sort.}

\item{checkRowIds}{Check if all rowIds in the covariates appear in the outcomes.}

\item{quiet}{If true, (warning) messages are surpressed.}

\item{overwrite}{Automatically overwrite if an index already exists?}
}
\value{
Nothing
}
\description{
\code{buildKnn} loads data from two ffdf objects, and inserts them into a KNN classifier.
}
\details{
These columns are expected in the outcome object:
\tabular{lll}{
  \verb{rowId}  	\tab(integer) \tab Row ID is used to link multiple covariates (x) to a single outcome (y) \cr
  \verb{y}    \tab(real) \tab The outcome variable \cr
}

These columns are expected in the covariates object:
\tabular{lll}{
  \verb{rowId}  	\tab(integer) \tab Row ID is used to link multiple covariates (x) to a single outcome (y) \cr
  \verb{covariateId}    \tab(integer) \tab A numeric identifier of a covariate  \cr
  \verb{covariateValue}    \tab(real) \tab The value of the specified covariate \cr
}

Note: If checkSorting is turned off, the covariate table should be sorted by rowId.
}

