% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlpInterface.R
\name{buildKnnFromPlpData}
\alias{buildKnnFromPlpData}
\title{Build a K-nearest neighbor (KNN) classifier from a plpData object}
\usage{
buildKnnFromPlpData(plpData, indexFolder, overWrite = TRUE,
  removeDropouts = TRUE, cohortId = NULL, outcomeId = NULL)
}
\arguments{
\item{plpData}{An object of type \code{plpData}.}

\item{indexFolder}{Path to a local folder where the KNN classifier index can be stored.}

\item{removeDropouts}{If TRUE subjects that do not have the full
observation window (i.e. are censored earlier) and do not have the
outcome are removed prior to fitting the model.}

\item{cohortId}{The ID of the specific cohort for which to fit a model.}

\item{outcomeId}{The ID of the specific outcome for which to fit a model.}

\item{overwrite}{Automatically overwrite if an index already exists?}
}
\description{
Build a K-nearest neighbor (KNN) classifier from a plpData object
}

